removed obsolete code and reformatted
